daily/20210324/example_cloud_functions/03service_account/requirements.txt
Flask==1.0.2
google-auth
google-api-python-client
